PW-5490 - Move both interface and implementation to a new directory
