+++ etc/webapi.xml
@@ -35,7 +35,7 @@
     </route>
 
     <route url="/V1/internal/guest-carts/:cartId/retrieve-adyen-payment-methods" method="POST">
-        <service class="Adyen\Payment\Api\InternalGuestAdyenPaymentMethodManagementInterface" method="handleInternalRequest"/>
+        <service class="Adyen\Payment\Api\Internal\InternalGuestAdyenPaymentMethodManagementInterface" method="handleInternalRequest"/>
         <resources>
             <resource ref="anonymous"/>
         </resources>
